Add initial Dockerfile for RunPod DwemerDistro template
Not tested yet. The idea is to dockerize DwemerDistro and put on Docker Hub.
Then users can deploy a Pod on RunPod with DwemerDistro with CUDA and xtts built in.

Then we need to make a Windows run script or .exe that will:
1) Start a Pod (use runpod CLI or runpod API with users credentials)
2) SSH tunnel to the Pod so that AIAgent mod can connect to Dwemer via localhost port
3) SSH tunnel other ports like webpage, postgres, whatever else needed

And next steps: fill in as much configuration in Docker DwemerDistro as possible:
1) Deepgram STT (free, but prompt for user key)
2) OR LLM (some well tested one)
3) XTTS (running in Docker DwemerDistro)
4) Deepinfra ITT (practically free, prompt for user key)
5) Basic prompts etc

No change for AIAgent .ini should be needed - will connect via localhost port that's SSH tunneled to RunPod while it lives