Add GPT-SoVITS .bat install script
The script downloads and installs GPT-SoVITS repo and pretrained models.
It installs miniconda, creates a conda env and runs install.sh from GPT-SoVITS.

NOTE: lots of GB downloaded and storage use.

After that, everything is installed and SoVITS API server can be run by activating conda and going:
 python api_v2.py -a 0.0.0.0 -p 9880 -c GPT_SoVITS/configs/tts_infer.yaml

then this connector can be used as client:
SOVIET_TTS_AS_XTTS_FASTAPI.php
which is included in this commit

you need to put the reference voice file in GPT_SoVITS. later on it will be voice ID concatenated with ".wav" to find the ref file